+++ src/main/java/jeff/command/Command.java
@@ -1,3 +1,13 @@
+package jeff.command;
+
+import jeff.main.JeffException;
+
+import jeff.storage.Storage;
+
+import jeff.task.TaskList;
+
+import jeff.ui.Ui;
+
 /**
  * Command class acts as a format for other Command classes.
  */
