+++ src/main/java/jeff/command/TodoCommand.java
@@ -1,3 +1,15 @@
+package jeff.command;
+
+import jeff.main.JeffException;
+
+import jeff.storage.Storage;
+
+import jeff.task.Task;
+import jeff.task.TaskList;
+import jeff.task.Todo;
+
+import jeff.ui.Ui;
+
 /**
  * TodoCommand class is a Command that contains instructions
  * to run when user wants to add a new Todo task.
