+++ src/main/java/jeff/command/UnmarkCommand.java
@@ -1,29 +1,39 @@
+package jeff.command;
+
+import jeff.main.JeffException;
+
+import jeff.storage.Storage;
+
+import jeff.task.TaskList;
+
+import jeff.ui.Ui;
+
 /**
- * MarkCommand class is a Command that contains instructions
- * to run when user wants to mark a specific task as done.
+ * UnmarkCommand class is a Command that contains instructions
+ * to run when user wants to mark a specific task as not done.
  */
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
 
     private int index;
 
     /**
-     * Constructor of MarkCommand.
+     * Constructor of UnmarkCommand.
      *
-     * @param body Position of the task to mark as done.
+     * @param body Position of the task to mark as not done.
      */
-    public MarkCommand(String body) {
+    public UnmarkCommand(String body) {
         this.index = Integer.parseInt(body) - 1;
     }
 
     /**
-     * Marks the Task as done according
+     * Marks the Task as not done according
      * to index given by the user if available.
      *
      * @param tasks TaskList containing all the tasks.
      * @param ui Ui class for invoking user feedback.
      * @param storage Storage class used to save files.
-     * @throws JeffException When index of task to mark as done is
-     *                       out of bounds or file cannot be saved.
+     * @throws JeffException When index of task to mark as not done
+     *                       is out of bounds or file cannot be saved.
      */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws JeffException {
@@ -31,12 +41,12 @@ public class MarkCommand extends Command {
             tasks.getAt(index);
         } catch (IndexOutOfBoundsException e) {
             int temp = index + 1;
-            throw new JeffException(" ☹ OOPS!!! The task you want to mark as done at index "
+            throw new JeffException(" ☹ OOPS!!! The task you want to mark as not done at index "
                     + temp + " is out of bounds,"
                     + "please double check the index number");
         }
-        tasks.mark(index);
-        ui.showMark(tasks.getString(index));
+        tasks.unmark(index);
+        ui.showUnmark(tasks.getString(index));
         storage.save(tasks);
     }
