+++ src/main/java/jeff/parser/Parser.java
@@ -1,3 +1,18 @@
+package jeff.parser;
+
+import jeff.main.JeffException;
+
+import jeff.command.ByeCommand;
+import jeff.command.Command;
+import jeff.command.DeadlineCommand;
+import jeff.command.DeleteCommand;
+import jeff.command.EventCommand;
+import jeff.command.HelpCommand;
+import jeff.command.ListCommand;
+import jeff.command.MarkCommand;
+import jeff.command.TodoCommand;
+import jeff.command.UnmarkCommand;
+
 /**
  * Parser class is used to parse the raw input from the user
  * and call the correct Command class while handling exceptions.
