+++ src/main/java/jeff/parser/TimeParse.java
@@ -1,3 +1,5 @@
+package jeff.parser;
+
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
