+++ src/main/java/jeff/task/Event.java
@@ -1,7 +1,14 @@
+package jeff.task;
+
+import jeff.main.JeffException;
+
+import jeff.parser.DateParse;
+import jeff.parser.TimeParse;
+
 /**
- * Deadline class is a task customised to store end date of deadlines.
+ * Event class is a task customised to store date of the Event.
  */
-public class Deadline extends Task {
+public class Event extends Task {
 
     protected String dateInfo;
     protected String originalDate;
@@ -9,13 +16,13 @@ public class Deadline extends Task {
     protected TimeParse time;
 
     /**
-     * Constructor for Deadline class.
+     * Constructor for Event class.
      *
-     * @param description Name of the Deadline.
+     * @param description Name of the Event.
      * @param dateInfo Information for which this task is due.
      * @throws JeffException When no available format is available to parse dateInfo.
      */
-    public Deadline(String description, String dateInfo) throws JeffException {
+    public Event(String description, String dateInfo) throws JeffException {
         super(description);
         this.originalDate = dateInfo;
         String[] str = dateInfo.split(" ", 2);
@@ -27,10 +34,10 @@ public class Deadline extends Task {
     /**
      * Returns the identity of this Task class.
      *
-     * @return D for Deadline.
+     * @return E for Event.
      */
     public String whatType() {
-        return "D";
+        return "E";
     }
 
     /**
@@ -44,12 +51,12 @@ public class Deadline extends Task {
     }
 
     /**
-     * toString method specific for Deadline class,
+     * toString method specific for Event class,
      * inherits toString() fromTask class while adding additional information.
-     * Like the type of task, [D], and date information.
+     * Like the type of task, [E], and date information.
      */
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + this.dateInfo + ")";
+        return "[E]" + super.toString() + " (at: " + this.dateInfo + ")";
     }
 }
