+++ src/main/java/jeff/task/Todo.java
@@ -1,3 +1,5 @@
+package jeff.task;
+
 /**
  * Todo class is a task customised to store the descriptions of task to do.
  */
